+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Smaller base images in Dockerfile. Fixes MODREP-24.
 * Upgrade golang from 1.23 to 1.24. Fixes MODREP-30.
 * Support validation of report URLs by reference to a whitelist of regexps in the configuration file. Fixes MODREP-4.
+* In JSON queries, ignore order-by elements when the fieldname is empty. Fixes MODREP-28.
+* When returning a report or JSON query response, the fields of the results are ordered as specified in query. Fixes MODREP-31.
 
 ## [1.3.1](https://github.com/folio-org/mod-reporting/tree/v1.3.1) (2025-03-12)